+++ src/SongGenre.cpp
@@ -9,7 +9,7 @@ int InstanceCounter<SongGenre>::s_maxCount;
 SongGenre::SongGenre(int id,
                      QString const& name,
                      int count,
-                     int duration,
+                     double duration,
                      QObject *parent)
     : ItemGroupBase(id, name, count, duration, parent) {
 }
